Add files via upload
This photo shows my (new) hardware, I made adapters for 20pin and 24pin GALS, it is easy to add any programming socket as the
adapter is separate from the programming part. The programming hardware can set VCC to 3V3 or 5V. The VPP can be set anywhere between 12.0V and 17.0 Volts and  ( a special) version of perlblast sets the correct values for the GAL. Also the voltages on the ZIF-socket in 'Off' mode are around 1.5V maximum.

As discussed, this LPT dongle works with the uss720 kernel module and creates a working /dev/parportx device. I took of the IEE1284 connector and replaced it by the 'standard' 25 pin female connector.